get hit dice from first class item
Instead of assuming the first sheet item is a class (which is not always the case if a user adds spells/items to their sheet first), attempt to find all class items and display their hit dice as a comma-separated list. If there are no class items, display the text "unknown" instead.